apps/server/requirements.txt
torch
torchvision
Pillow
onnxruntime
